% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2stanR.R
\name{df2stan}
\alias{df2stan}
\title{df2stan}
\usage{
df2stan(data, include_missing = FALSE, is_log_bsi = FALSE)
}
\arguments{
\item{data}{A data.frame. An example can be found by simulating data with `sim_counts` and looking at the wide-format}

\item{include_missing}{A boolean value on whether to include or exclude missing data (if included, NA will be replaced with the mean of the relevant data)}

\item{is_log_bsi}{A boolean value determing if total BSI is on the log scale or not}
}
\description{
A function to transform data.frame in to a stan data list ready for the models described here
}
