% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gather_counts}
\alias{gather_counts}
\title{gather_counts}
\usage{
gather_counts(data)
}
\arguments{
\item{data}{A data.frame as supplied to df2stan}
}
\description{
A function to gather all the counts for the different species in a table
}
