% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_prediction.R
\name{posterior_predict}
\alias{posterior_predict}
\title{Posterior predictive check}
\usage{
posterior_predict(fit, draws, seed)
}
\arguments{
\item{fit:}{a stanfit object fit by one of the models in this packages}

\item{new_data:}{a data.frame containing the new data to estimate}

\item{draws:}{how many draws from the posterior}

\item{seed:}{seed the random generator}
}
\description{
Posterior predictive check
}
