% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_models_ni}
\alias{run_models_ni}
\title{Simple Bayesian models with no missing data imputation}
\usage{
run_models_ni(data, model, ...)
}
\arguments{
\item{data}{A data.frame with some data in the same format as ouputted by `sim_counts`}

\item{model}{A string specifying which model to run}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
An object of class `stanfit` returned by `rstan::sampling`
}
\description{
Simple Bayesian models with no missing data imputation
}
