% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_counts.R
\name{sim_counts}
\alias{sim_counts}
\title{A function to simulate counts}
\usage{
sim_counts(disease_param, inst_param, years, seed = NULL)
}
\arguments{
\item{disease_param}{A data frame with disease parameters}

\item{inst_param}{A data frame with institutional parameters}

\item{years}{A vector of years}

\item{seed}{An integer to set the seed.}
}
\value{
A list with two elements:
          `long_sim`: a tibble with seven columns:
          (1) disease_type;
          (2) inst_id;
          (3) total_bsi: the reported bsi by the institution (-99 if missing);
          (4) count_r: the reported count by the institution (-99 if missing);
          (5) bsi: the actual bsi (equal to bsi_r if reporting is TRUE);
          (6) count: the actual count (equal to count_r if reporting is TRUE);
          (7) pi_dy: the year-specific disease rate
          `wide_sim`: a tibble the more closely resembles what someone's data might look like:
          (1) labcode;
          (2) year;
          (3) total_bsi; # data reported by a lab
          (4) one or more columns with counts of data for each species as reported by lab
}
\description{
A function to simulate counts
}
