% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sims2data}
\alias{sims2data}
\title{sims2data}
\usage{
sims2data(sims)
}
\arguments{
\item{table}{The output from sim_counts}
}
\description{
Transform sim_counts output to a data table that might resemble what people might actually have.
}
